Device: Move FileIncluder into the source file
Doesn't need to be in the header.
